requirements.txt
langchain
langchain-google-genai
langchain-community
chromadb
faiss-cpu
reportlab
gradio
python-dotenv
langfuse
pypdf
